+++ MaterialDidatico/src/main/webapp/app/disciplina/form.xhtml
@@ -23,6 +23,19 @@ and open the template in the editor.
                         <p:focus for="nome"/>
                     </p:inputText>
                     
+                    <p:outputLabel value="Curso:" for="curso"/>
+                    <p:panel>
+                        <p:autoComplete id="curso" maxResults="5" dropdown="true"
+                                        completeMethod="#{cursoControle.AutoCompletePorNome}" 
+                                        value="#{disciplinaControle.entidade.curso}"
+                                        converter="#{cursoControle.converter()}"
+                                        var="item"
+                                        itemLabel="#{item.nome}"
+                                        itemValue="#{item}"/>
+                        <p:commandButton icon="fa fa-plus" style="font-size: small;"
+                                         action="#{disciplinaControle.cadastroCurso()}"
+                                         ajax="false" immediate="true"/>
+                    </p:panel>
                 </h:panelGrid>
 
                 <f:facet name="footer">
